+++ dwds/CHANGELOG.md
@@ -2,8 +2,7 @@
 
 - Update SDK constraint to `>=2.14.0 <3.0.0`
 - Depend on `vm_service` version `7.3.0`.
-- Add metrics for API on the critical path for DevTools IPL.
-- Add metrics for total DevTools IPL.
+- Add metrics measuring DevTools Initial Page Load time.
 
 ## 11.2.3